Added snowflake adapter
