+++ core/adapters/snowflake.ts
@@ -40,7 +40,8 @@ export class SnowflakeAdapter extends Adapter implements IAdapter {
             this.insertInto(
               table.target,
               tableMetadata.fields.map(f => f.name),
-              this.where(table.incrementalQuery || table.query, table.where)
+              this.where(table.incrementalQuery || table.query, table.where),
+              table.uniqueKey
             )
           )
         );
@@ -79,5 +80,25 @@ export class SnowflakeAdapter extends Adapter implements IAdapter {
     return `create or replace ${this.baseTableType(table.type || "table")} ${this.resolveTarget(
       table.target
     )} as ${table.query}`;
+  }
+
+  public insertInto(
+    target: dataform.ITarget,
+    columns: string[],
+    query: string,
+    uniqueKey: string[]
+  ) {
+    return `
+merge into ${this.resolveTarget(target)} T
+using (${query}) S
+on ${
+      uniqueKey && uniqueKey.length > 0
+        ? uniqueKey.map(uk => `T.${uk} = S.${uk}`).join(` and `)
+        : `false`
+    }
+when matched then
+  update set ${columns.map(c => `${c} = S.${c}`).join(",")}
+when not matched then
+  insert (${columns.join(",")}) values (${columns.join(",")})`;
   }
 }
