+++ tests/integration/snowflake_project/definitions/example_incremental_merge.sqlx
@@ -1,0 +1,15 @@
+WITH example_data AS (
+SELECT ts AS ts, id_1 AS id_1, id_2 AS id_2, val AS val
+FROM 
+  (SELECT 1 AS ts,	21	AS id_1, 31 AS id_2, 'original' AS val) UNION ALL
+  (SELECT 2 AS ts,	22	AS id_1, 32 AS id_2, 'original' AS val) UNION ALL
+  (SELECT 3 AS ts,	21	AS id_1, 31 AS id_2, 'new' AS val) UNION ALL
+  (SELECT 4 AS ts,	22	AS id_1, 32 AS id_2, 'new' AS val)
+)
+
+config { type: "incremental", uniqueKey: ["id_1", "id_2"] }
+
+SELECT ts, id_1, id_2, val
+FROM example_data
+
+${ when(incremental(), `WHERE ts > 2`, `WHERE ts <= 2`) }
